refactor: 게이머과 딜러의 게임 결과 판단 기능 딜러에서 GameResult 객체로 책임 이동
