+++ docs/README.md
@@ -35,3 +35,4 @@
     - [x] 딜러는 자신의 점수를 계산할 수 있다.
     - [x] 플레이어를 받아서 승패를 판단할 수 있다.
     - [x] 딜러는 점수가 16이하면 카드를 추가할 수 있다.
+- [ ]
